+++ dev/user.clj
@@ -4,10 +4,14 @@
     [byte-streams :as bytes :refer [bytes=]]
     (blocks
       [core :as block]
-      [data :as data])
+      [data :as data]
+      [store :as store])
     (blocks.store
-      [memory :refer [memory-block-store]]
+      [buffer :refer [buffer-block-store]]
+      [cache :refer [caching-block-store]]
       [file :refer [file-block-store]]
+      [memory :refer [memory-block-store]]
+      [replica :refer [replica-block-store]]
       [tests :as tests])
     [clojure.java.io :as io]
     [clojure.repl :refer :all]
@@ -15,7 +19,8 @@
     [clojure.test.check :as check]
     [clojure.test.check.generators :as gen]
     [clojure.test.check.properties :as prop]
-    [multihash.core :as multihash])
+    [multihash.core :as multihash]
+    [multihash.digest :as digest])
   (:import
     blocks.data.Block
     multihash.core.Multihash))