+++ src/blocks/store/cache.clj
@@ -1,24 +1,19 @@
 (ns blocks.store.cache
-  "Logical block storage backed by two other stores, a _primary store_ and a
-  _cache_. Blocks are added to the cache on reads and writes, and evicted with
-  a least-recently-used strategy to keep the cache under a certain total size.
+  "Cache stores provide logical block storage backed by two other stores, a
+  _primary store_ and a _cache_. Blocks are added to the cache on reads and
+  writes, and evicted with a least-recently-used strategy to keep the cache
+  under a certain total size. Operations on this store will prefer to look up
+  blocks in the cache, and fall back to the primary store when not available.
 
-  Operations on this store will prefer to look up blocks in the cache, and fall
-  back to the primary store when not available."
+  Because the caching logic runs locally, the backing cache storage should not
+  be shared among multiple concurrent processes."
   (:require
     [blocks.core :as block]
     [blocks.store :as store]
-    [blocks.store.util :as util]
+    [blocks.summary :as sum]
     [clojure.data.priority-map :refer [priority-map]]
     [clojure.tools.logging :as log]
     [com.stuartsierra.component :as component]))
-
-
-; TODO: use a generic sorted-kv interface to persist state.
-; keys would look like:
-; [:priorities tick size] -> block id
-; [:total-size] -> size
-; [:tick] -> tick counter
 
 
 (defn- scan-state
@@ -27,7 +22,9 @@
   [store]
   (reduce
     (fn [state block]
-      (let [tick (or (some-> ^java.util.Date (:stored-at block) (.getTime)) 0)]
+      ; TODO: could do something clever with the :stored-at metadata, but this
+      ; is probably sufficient.
+      (let [tick 0]
         (-> state
             (update :priorities assoc (:id block) [tick (:size block)])
             (update :total-size + (:size block))
@@ -53,23 +50,24 @@
 
 (defn reap!
   "Given a target amount of space to free and a cache store, deletes blocks from
-  the cache to free up the desired amount of space. Returns a vector of the
+  the cache to free up the desired amount of space. Returns a summary of the
   deleted entries."
   [store target-free]
   (let [{:keys [cache state size-limit]} store]
-    (loop [deleted []]
-      (let [{:keys [priorities total-size]} @state]
-        (if (and (< (- size-limit total-size) target-free)
-                 (not (empty? priorities)))
-          ; Need to delete the next block.
-          (let [[id [tick size]] (peek priorities)]
-            (block/delete! cache id)
-            (swap! state assoc
-                   :total-size (- total-size size)
-                   :priorities (pop priorities))
-            (recur (conj deleted {:id id, :tick tick, :size size})))
-          ; Enough free space, or no more blocks to delete.
-          deleted)))))
+    (locking store
+      (loop [deleted (sum/init)]
+        (let [{:keys [priorities total-size]} @state]
+          (if (and (< (- size-limit total-size) target-free)
+                   (not (empty? priorities)))
+            ; Need to delete the next block.
+            (let [[id [tick size]] (peek priorities)]
+              (block/delete! cache id)
+              (swap! state #(-> %
+                                (update :total-size - size)
+                                (update :priorities pop)))
+              (recur (sum/update deleted {:id id, :size size})))
+            ; Enough free space, or no more blocks to delete.
+            deleted))))))
 
 
 (defn- maybe-cache!
@@ -84,11 +82,10 @@
       ; Store the block and update cache state.
       (when-let [cached (block/put! cache block)]
         (swap! (:state store)
-               (fn update-state
-                 [{:keys [priorities total-size tick]}]
-                 {:priorities (assoc priorities (:id cached) [tick (:size cached)])
-                  :total-size (+ total-size (:size cached))
-                  :tick (inc tick)}))
+               #(-> %
+                    (update :priorities assoc (:id cached) [(:tick %) (:size cached)])
+                    (update :total-size + (:size cached))
+                    (update :tick inc)))
         cached))))
 
 
@@ -113,7 +110,7 @@
         (let [initial-state (scan-state cache)
               cached-bytes (:total-size initial-state)]
           (reset! state initial-state)
-          (when (and cached-bytes (pos? cached-bytes))
+          (when (pos? cached-bytes)
             (log/infof "Cache has %d bytes in %d blocks"
                        (:total-size initial-state)
                        (count (:priorities initial-state))))
@@ -137,9 +134,9 @@
   (-list
     [this opts]
     (ensure-initialized! this)
-    (util/select-stats
+    (store/select-stats
       opts
-      (util/merge-block-lists
+      (store/merge-block-lists
         (store/-list cache   opts)
         (store/-list primary opts))))
 
@@ -158,7 +155,7 @@
     (ensure-initialized! this)
     (when-let [id (:id block)]
       (let [cached (maybe-cache! this block)
-            preferred (util/preferred-copy cached block)
+            preferred (store/preferred-copy cached block)
             stored (store/-put! primary preferred)]
         (or cached stored))))
 
@@ -172,6 +169,9 @@
 
 
 ;; ## Constructors
+
+(store/privatize-constructors! CachingBlockStore)
+
 
 (defn caching-block-store
   "Creates a new logical block store which will use one block store to cache
@@ -198,8 +198,3 @@
     (assoc opts
            :size-limit size-limit
            :state (atom nil))))
-
-
-;; Remove automatic constructor functions.
-(ns-unmap *ns* '->CachingBlockStore)
-(ns-unmap *ns* 'map->CachingBlockStore)