+++ src/blocks/store/file.clj
@@ -1,6 +1,7 @@
 (ns blocks.store.file
   "Block storage backed by files in nested directories. Each block is stored in
-  a separate file.
+  a separate file. File block stores may be constructed using the
+  `file://<path-to-root>` URI form.
 
   In many filesystems, performance degrades as the number of files in a
   directory grows. In order to reduce this impact and make navigating the
@@ -25,7 +26,6 @@
       [core :as block]
       [data :as data]
       [store :as store])
-    [blocks.store.util :as util]
     [clojure.java.io :as io]
     [clojure.string :as str]
     [clojure.tools.logging :as log]
@@ -118,11 +118,11 @@
     (some->
       file
       (.getPath)
-      (util/check #(.startsWith ^String % root)
+      (store/check #(.startsWith ^String % root)
         (log/warnf "File %s is not a child of root directory %s" file root))
       (subs (inc (count root)))
       (str/replace "/" "")
-      (util/check hex/valid?
+      (store/check hex/valid?
         (log/warnf "File %s did not form valid hex entry: %s" file value))
       (multihash/decode))))
 
@@ -167,7 +167,7 @@
     [this opts]
     (->> (find-files root (:after opts))
          (keep #(block-stats (file->id root %) %))
-         (util/select-stats opts)))
+         (store/select-stats opts)))
 
 
   (-get
@@ -191,31 +191,24 @@
 
   (-delete!
     [this id]
-    (or (when-block id
-          (locking this
-            (.delete file)))
-        false))
+    (boolean (when-block id
+               (locking this
+                 (.delete file)))))
 
 
-  store/BlockEnumerator
+  store/ErasableStore
 
-  (enumerate
+  (-erase!
     [this]
-    (->> (.listFiles root)
-         (mapcat #(.listFiles ^File %))
-         (map #(file->block (file->id root %) %)))))
-
-
-(defn erase!
-  "Clears all contents of the file store by recursively deleting the root
-  directory."
-  [store]
-  (locking store
-    (rm-r (:root store))))
+    (locking this
+      (rm-r root))))
 
 
 
 ;; ## Constructors
+
+(store/privatize-constructors! FileBlockStore)
+
 
 (defn file-block-store
   "Creates a new local file-based block store."
@@ -226,13 +219,8 @@
 
 (defmethod store/initialize "file"
   [location]
-  (let [uri (util/parse-uri location)]
+  (let [uri (store/parse-uri location)]
     (file-block-store
       (if (:host uri)
         (io/file (:host uri) (subs (:path uri) 1))
         (io/file (:path uri))))))
-
-
-;; Remove automatic constructor functions.
-(ns-unmap *ns* '->FileBlockStore)
-(ns-unmap *ns* 'map->FileBlockStore)