+++ test/blocks/bytes_test.clj
@@ -6,6 +6,12 @@
     blocks.data.PersistentBytes))
 
 
+(defn- ->pb
+  "Construct a new `PersistentBytes` value containing the given byte data."
+  [& data]
+  (PersistentBytes/wrap (byte-array data)))
+
+
 (deftest persistent-bytes-construction
   (testing "wrap does not duplicate the array"
     (let [data (byte-array 3)
@@ -48,3 +54,28 @@
     (is (= 97 (nth pb 1)))
     (is (thrown? Exception (nth pb 5)))
     (is (= ::not-found (nth pb 5 ::not-found)))))
+
+
+(deftest byte-comparison
+  (testing "equal arrays"
+    (is (zero? (compare (->pb) (->pb))))
+    (is (zero? (compare (->pb 1 2 3) (->pb 1 2 3)))))
+  (testing "equal prefixes"
+    (is (neg? (compare (->pb 1 2 3)
+                       (->pb 1 2 3 4))))
+    (is (pos? (compare (->pb 1 2 3 4)
+                       (->pb 1 2 3)))))
+  (testing "order-before"
+    (is (neg? (compare (->pb 1 2 3)
+                       (->pb 1 2 4))))
+    (is (neg? (compare (->pb 1 2 3)
+                       (->pb 1 3 2))))
+    (is (neg? (compare (->pb 0 2 3 4)
+                       (->pb 1 3 2 1))))
+  (testing "order-after"
+    (is (pos? (compare (->pb 1 2 4)
+                       (->pb 1 2 3))))
+    (is (pos? (compare (->pb 1 3 2)
+                       (->pb 1 2 3))))
+    (is (pos? (compare (->pb 1 3 2 1)
+                       (->pb 0 2 3 4)))))))